requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-pptx
pypdf
python-docx
google-generativeai
unstructured
tiktoken
